+++ CHANGES.rst
@@ -30,6 +30,7 @@ Improvements
   refreshing identity information
 - Show more verbose error when email validation fails during event
   registration (:issue:`4177`)
+- Add link to external map in room details view (:issue:`4146`)
 
 Bugfixes
 ^^^^^^^^